google: reform Geocoding/FindPlace scripts
Apply a variety of reforms for the geocoding and place-finding
algorithms, among which:

(1) Adapt to using placename_orig, endr_orig, city and UF as keys in the
R scripts, which means all of these keys need to be available in all
exported CSVs.

(2) Adapt to generating multiple files with slightly higher scale by
respecting names.

(3) Make the filter scripts tolerate ID-less legacy JSON files.

This change comes with some caveats: first, data generated after these
changes is not fully compatible with legacy data generated before them
(e.g.  the endr_orig key generated by find_place.py is different than it
used to be). Second, there are some admittedly irrational
inconsistencies (endr_orig in find_place.py does not correspond to
endr_orig in geocode.py). This generates some complexity on the R side,
but it was the best compromise I found to accomodate the legacy stuff
given the changes that came to be required.