google_matcher: match across different years
Google Matcher has an erratic history. We moved from a format where we
matched by address, into a format where we matched by ID, and now we're
moving back to a format where we match by state, city, and address. This
is because this is the only way to match across different years of the
polling stations database.

This change came along with a refactor of the matcher to use more
dplyr-esque syntax.