create_datasets: make Malha 2019 actually parseable
Unsurprisingly, the IBGE has delivered the Malha 2019 with some
shapefile errors -- particularly, linestrings with a single point.
However, removing these few erroneous features is extremely
time-consuming, so we had to come up with some optimizations to do so.
This is the result.

Basically, we only run corrections when st_centroid() fails; and we
filter out duplicate streets or streets without businesses.

In this commit, we also optimize merging these shapefiles by using
data.table::rbindlist() instead of purrr::map_dfr(). This is a *lot*
faster.